+++ src/main/java/com/arangodb/internal/ArangoDefaults.java
@@ -29,7 +29,8 @@ import com.arangodb.entity.LoadBalancingStrategy;
  */
 public final class ArangoDefaults {
 
-	private ArangoDefaults() {
+
+    private ArangoDefaults() {
 		super();
 	}
 
@@ -41,6 +42,7 @@ public final class ArangoDefaults {
 	public static final Integer DEFAULT_TIMEOUT = 0;
 	public static final String DEFAULT_USER = "root";
 	public static final Boolean DEFAULT_USE_SSL = false;
+	public static final Boolean DEFAULT_HTTP_RESEND_COOKIES = true;
 	public static final int HEADER_SIZE = INTEGER_BYTES + INTEGER_BYTES + LONG_BYTES + LONG_BYTES;
 	public static final int CHUNK_DEFAULT_CONTENT_SIZE = 30000;
 	public static final int MAX_CONNECTIONS_VST_DEFAULT = 1;